添加了初始化数据的方法

+++ src/main/java/me/jcala/jmooc/entity/Chapter.java
@@ -12,7 +12,7 @@ public class Chapter {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;//id
+    private int id;//id
 
     @Column(nullable = false,length = 40)
     private String name;//章节名称
@@ -32,11 +32,11 @@ public class Chapter {
     }
 
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
