+++ src/main/java/me/jcala/jmooc/entity/Exercise.java
@@ -16,7 +16,7 @@ public class Exercise {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;//id
+    private int id;//id
 
     @Column(nullable = false,length = 40)
     private String name;//习题名称
@@ -42,11 +42,11 @@ public class Exercise {
     }
 
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
